convert to reusable error package
Change package name from 'main' to 'error' in source and test files to make
the error injection and extraction logic reusable as a package. Update README
examples to use 'error' package. Remove main.go and clarify in README that
the code is intended to be copied rather than imported as a library.

Signed-off-by: ayn <ayn.khatib@gmail.com>